+++ packages/zowe-explorer-api/src/profiles/ProfilesCache.ts
@@ -162,7 +162,6 @@ export class ProfilesCache {
 
     public async refresh(apiRegister?: ZoweExplorerApi.IApiRegisterClient): Promise<void> {
         this.allProfiles = [];
-        let tmpAllProfiles: zowe.imperative.IProfileLoaded[] = [];
         this.allTypes = [];
         let mProfileInfo: zowe.imperative.ProfileInfo;
         try {
@@ -170,6 +169,7 @@ export class ProfilesCache {
             const allTypes = this.getAllProfileTypes(apiRegister.registeredApiTypes());
             allTypes.push("base");
             for (const type of allTypes) {
+                const tmpAllProfiles: zowe.imperative.IProfileLoaded[] = [];
                 // Step 1: Get all profiles for each registered type
                 const profilesForType = mProfileInfo
                     .getAllProfiles(type)
@@ -190,7 +190,6 @@ export class ProfilesCache {
                     }
                     this.allProfiles.push(...tmpAllProfiles);
                     this.profilesByType.set(type, tmpAllProfiles);
-                    tmpAllProfiles = [];
                 }
                 this.allTypes.push(type);
             }
@@ -289,6 +288,20 @@ export class ProfilesCache {
             }
         }
         return profByType;
+    }
+
+    /**
+     * get array of IProfileLoaded for all profiles
+     * @returns IProfileLoaded[]
+     */
+    public async fetchAllProfiles(): Promise<zowe.imperative.IProfileLoaded[]> {
+        const profiles: zowe.imperative.IProfileLoaded[] = [];
+        for (const type of this.allTypes) {
+            const profsByType = await this.fetchAllProfilesByType(type);
+            profiles.push(...profsByType);
+        }
+        this.allProfiles = profiles;
+        return profiles;
     }
 
     /**